TPT3-159: activeHealthInterventionPlans moved to ChildClientCoreServiceImpl
