+++ legacy-data-access-services/src/main/java/gov/ca/cwds/cms/data/access/dto/ClientEntityAwareDTO.java
@@ -4,15 +4,12 @@ import gov.ca.cwds.data.legacy.cms.entity.Client;
 import gov.ca.cwds.data.legacy.cms.entity.ClientScpEthnicity;
 import gov.ca.cwds.data.legacy.cms.entity.DeliveredService;
 import gov.ca.cwds.data.legacy.cms.entity.MedicalEligibilityApplication;
-import gov.ca.cwds.data.legacy.cms.entity.HealthInterventionPlan;
-
 import java.util.ArrayList;
 import java.util.List;
 
 public class ClientEntityAwareDTO extends BaseEntityAwareDTO<Client> {
   private List<ClientScpEthnicity> clientScpEthnicities = new ArrayList<>();
   private List<MedicalEligibilityApplication> medicalEligibilityApplications = new ArrayList<>();
-  private List<HealthInterventionPlan> activeHealthInterventionPlans = new ArrayList<>();
   private List<DeliveredService> deliveredService;
 
   public List<ClientScpEthnicity> getClientScpEthnicities() {
@@ -30,15 +27,6 @@ public class ClientEntityAwareDTO extends BaseEntityAwareDTO<Client> {
   public void setMedicalEligibilityApplications(
       List<MedicalEligibilityApplication> medicalEligibilityApplications) {
     this.medicalEligibilityApplications = medicalEligibilityApplications;
-  }
-
-  public List<HealthInterventionPlan> getActiveHealthInterventionPlans() {
-    return activeHealthInterventionPlans;
-  }
-
-  public void setActiveHealthInterventionPlans(
-      List<HealthInterventionPlan> activeHealthInterventionPlans) {
-    this.activeHealthInterventionPlans = activeHealthInterventionPlans;
   }
 
   public List<DeliveredService> getDeliveredService() {
