+++ legacy-data-access-services/src/main/java/gov/ca/cwds/cms/data/access/service/impl/ChildClientCoreServiceImpl.java
@@ -4,9 +4,11 @@ import com.google.inject.Inject;
 import gov.ca.cwds.cms.data.access.dto.ChildClientEntityAwareDTO;
 import gov.ca.cwds.cms.data.access.service.ChildClientCoreService;
 import gov.ca.cwds.data.legacy.cms.dao.HealthInterventionPlanDao;
+import gov.ca.cwds.data.legacy.cms.dao.ParentalRightsTerminationDao;
 import gov.ca.cwds.data.legacy.cms.entity.ChildClient;
 import gov.ca.cwds.data.legacy.cms.entity.FCEligibility;
 import gov.ca.cwds.data.legacy.cms.entity.HealthInterventionPlan;
+import gov.ca.cwds.data.legacy.cms.entity.ParentalRightsTermination;
 import java.util.List;
 
 /**
@@ -17,18 +19,27 @@ public class ChildClientCoreServiceImpl extends
 
   @Inject
   private HealthInterventionPlanDao healthInterventionPlanDao;
+  @Inject
+  private ParentalRightsTerminationDao parentalRightsTerminationDao;
 
   protected void enrichClientEntityAwareDto(ChildClientEntityAwareDTO clientEntityAwareDTO) {
+
     ChildClient childClient = (ChildClient) clientEntityAwareDTO.getEntity();
+    final String childClientId = childClient.getIdentifier();
+
     if (childClient.getAfdcFcEligibilityIndicatorVar()) {
       List<FCEligibility> fcEligibilities =
-          getFcEligibilityDao().findByChildClientId(childClient.getIdentifier());
+          getFcEligibilityDao().findByChildClientId(childClientId);
       clientEntityAwareDTO.setFcEligibilities(fcEligibilities);
     }
 
     List<HealthInterventionPlan> activeHealthInterventionPlans =
-        healthInterventionPlanDao.getActiveHealthInterventionPlans(childClient.getIdentifier());
+        healthInterventionPlanDao.getActiveHealthInterventionPlans(childClientId);
     clientEntityAwareDTO.setActiveHealthInterventionPlans(activeHealthInterventionPlans);
+
+    List<ParentalRightsTermination> parentalRightsTerminations =
+        parentalRightsTerminationDao.getParentalRightsTerminationsByChildClientId(childClientId);
+    clientEntityAwareDTO.setParentalRightsTerminations(parentalRightsTerminations);
   }
 
   public void setHealthInterventionPlanDao(
